fix(actions): serialize concurrent world-time edits per entry
Two Saves during one in-flight write both read the pre-edit seconds and both dispatch, minting two actionIds and so two CTRL-Z steps for one edit — reproduced in the harness, which yields 2 deltas without the lock. The lock has to wrap the read, so unlike promoteStagedEntity it cannot be an opt-in inside applyDeltaAction; withKeyLock moves to its own module and the key names this action rather than updateStoryEntryMetadata, whose pipeline dispatchers never pass through here.
